// web/src/generated/api-v1/src/models/LiveActivityContentResult.ts
export type LiveActivityContentResult = { type: 'actionItemNotFound' } & ActionItemNotFoundResult | { type: 'contentAvailable' } & ContentAvailableResult | { type: 'contentUnavailable' } & ContentUnavailableResult | { type: 'newerActivityObserved' } & NewerActivityObservedResult | { type: 'staleActivityVersion' } & LiveStaleActivityVersionResult;